cxlinux/CortexLinuxcom-Website/cxlinux/cx-ops/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cx-ops"
version = "0.1.0"
description = "Operations toolkit for Cortex Linux"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Cortex Linux Team", email = "team@cortexlinux.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all-connectors = [
    "openai>=1.6.0",
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
cx-ops = "cx_ops.cli:app"

[project.urls]
Homepage = "https://github.com/cortexlinux/cx-ops"
Repository = "https://github.com/cortexlinux/cx-ops"
Documentation = "https://docs.cortexlinux.com/ops"

[tool.setuptools.packages.find]
where = ["."]
include = ["cx_ops*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
